Initial commit As you can tell, this project started from the HelloIOIO example. On initial commit, the IOIO board has 5 digital inputs for doors status and 2 digital outputs to control garage doors.
Right now this is an Android Application but the plan is to make it a service and have a "client" application that connects to it remotely either on the home WiFi or remotely (using Firebase?).

I can hard-code my phone number in and when in Alarm mode I get a text message if any door opens.

Everything really is hard-coded in this first revision, the text of the door names, the recorded announcement clips to play, the actions to take in each mode, etc.  Ideally these would all be configurable in the app.